+++ jobs/CHANGELOG.rst
@@ -8,7 +8,10 @@ This document describes changes between each past release.
 ------------------
 
 - Add ability to configure cache folder via environment variable ``CACHE_FOLDER``
+
+- Fix support of some funnelcake archives (fixes #287)
 - Skip very old RC release with parenthesis in filenames (fixes #288)
+
 
 0.6.0 (2017-10-10)
 ------------------